requirements.txt
flask
flask-socketio
flask-sqlalchemy
eventlet              
pycryptodome          
Pillow                
pywavelets            
numpy                 
pyngrok               
python-dotenv         